requirements.txt
# Core dependencies
gradio[mcp]==5.32.1
huggingface-hub>=0.17.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# Analytics and ML dependencies
scikit-learn>=1.3.0
statsmodels>=0.14.0
matplotlib>=3.7.0
plotly>=5.17.0

# Database dependencies
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0

# MCP (Model Context Protocol) related
mcp>=0.1.0

# API integrations
requests>=2.28.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Database integration
psycopg2-binary>=2.9.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
watchdog>=3.0.0
python-dotenv>=1.0.0

# Optional: for advanced features
accelerate>=0.20.0
datasets>=2.14.0
tokenizers>=0.13.0
Pillow>=9.5.0
opencv-python>=4.7.0

vertexai
